Move 'walletIdFromPublicId' to wallets index
